requirements.txt
attrs==22.1.0
environ-config==22.1.0
paho-mqtt==1.6.1
prometheus-client==0.14.1
pyserial==3.5
pysolarmanv5==2.4.0
uModbus==1.0.4
